Move MockConsumer to separate file so can be used in multiple test suites
